backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.10
alembic==1.16.5
pydantic-settings==2.11.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
redis==6.4.0
minio==7.2.18
python-multipart==0.0.20
email-validator==2.3.0
pytest==8.4.2
httpx==0.28.1
openpyxl==3.1.5
